# src/main/java/com/sprint/mission/discodeit/dto/user/UserDto.java
package com.sprint.mission.discodeit.dto.user;

import com.sprint.mission.discodeit.entity.User;
import com.sprint.mission.discodeit.entity.UserStatus;

import java.time.Instant;
import java.util.UUID;

public record UserDto(
        UUID id,
        Instant createdAt,
        Instant updatedAt,
        String username,
        String email,
        UUID profileId,
        boolean online

) {
    // User랑 DTO 호환용 from 메소드
    public static UserDto from(User user, UserStatus status) {
        return new UserDto(user.getId(),
                user.getCreatedAt(),
                user.getUpdatedAt(),
                user.getUsername(),
                user.getEmail(),
                user.getProfileId(),
                status.isUserOnline());
    }
}